Added some validations

+++ install.sh
@@ -7,7 +7,7 @@ echo "Creating root directory..."
 sudo mkdir -p /var/www/example.com/public_html
 
 echo "Setting up permissions..."
-nginx_user=$(grep user /etc/nginx/nginx.conf | awk '{print $2}' | sed 's/\;//g')
+nginx_user=$(grep user /etc/nginx/nginx.conf | head -1 | awk '{print $2}' | sed 's/\;//g')
 sudo chown -R $nginx_user:$nginx_user /var/www/example.com/public_html
 sudo chmod 755 /var/www
 
@@ -27,7 +27,12 @@ fi
 
 sudo cp files/example.com /etc/nginx/sites-available/example.com
 sudo ln -s /etc/nginx/sites-available/example.com /etc/nginx/sites-enabled/example.com
-sudo rm /etc/nginx/sites-enabled/default
+
+if [ -f '/etc/nginx/sites-enabled/default' ]
+then
+    sudo rm /etc/nginx/sites-enabled/default
+fi
+
 
 echo "Restarting nginx..."
 sudo vs_service --restart nginx
